refactor: remove MonoGame from the desktop runtime
so long, partner 😌
