+++ Directory.Packages.props
@@ -2,7 +2,6 @@
 
   <PropertyGroup>
     <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
-    <MonoGameVersion>3.8.5.1</MonoGameVersion>
     <SkiaSharpVersion>4.151.2</SkiaSharpVersion>
     <SDL3Version>3.4.16</SDL3Version>
     <SDL3MixerVersion>3.2.4.11</SDL3MixerVersion>
@@ -10,13 +9,6 @@
 
   <!-- Game -->
   <ItemGroup>
-    <PackageVersion Include="FontStashSharp.MonoGame" Version="1.5.8" />
-    <PackageVersion Include="Microsoft.NETCore.Jit" Version="1.0.12" />
-    <PackageVersion Include="MonoGame.Framework.Native" Version="$(MonoGameVersion)" />
-    <PackageVersion Include="MonoGame.Framework.DesktopGL" Version="$(MonoGameVersion)" />
-    <PackageVersion Include="MonoGame.Runtime.Windows.Vulkan" Version="$(MonoGameVersion)" />
-    <PackageVersion Include="MonoGame.Runtime.Mac.Vulkan" Version="$(MonoGameVersion)" />
-    <PackageVersion Include="MonoGame.Runtime.Linux.Vulkan" Version="$(MonoGameVersion)" />
     <PackageVersion Include="FFmpeg.AutoGen" Version="9.0.1.1" />
     <PackageVersion Include="SkiaSharp" Version="$(SkiaSharpVersion)" />
     <PackageVersion Include="SkiaSharp.NativeAssets.WebAssembly" Version="$(SkiaSharpVersion)" />
@@ -32,15 +24,6 @@
     <PackageVersion Include="SDL3-CS.MacOS.Mixer" Version="$(SDL3MixerVersion)" />
   </ItemGroup>
 
-  <!-- Content builder -->
-  <ItemGroup>
-    <PackageVersion Include="MonoGame.Framework.Content.Pipeline" Version="$(MonoGameVersion)" />
-    <PackageVersion Include="MonoGame.Tool.Basisu" Version="2.0.2.1" />
-    <PackageVersion Include="MonoGame.Tool.Crunch" Version="1.0.4.7" />
-    <PackageVersion Include="MonoGame.Tool.Dxc" Version="1.8.2505.11" />
-    <PackageVersion Include="MonoGame.Tool.FFmpeg" Version="7.0.0.10" />
-    <PackageVersion Include="MonoGame.Tool.FFprobe" Version="7.0.0.10" />
-  </ItemGroup>
 
   <!-- Tests -->
   <ItemGroup>
