--- alsoft.ini
@@ -1,2 +1,0 @@
-[general]
-hrtf = false
