+++ src/CutTheRopeDX.Browser/Browser/GameLoop.cs
@@ -121,7 +121,6 @@ namespace CutTheRopeDX.Browser
             int steps = 0;
             while (_accumulator >= StepSeconds && steps < MaxCatchUpSteps)
             {
-                SoundMgr.Update(TimeSpan.FromSeconds(StepSeconds));
                 // Every key read happens inside a step, and the step ends by clearing the
                 // presses it consumed. Core asks "was this pressed since I last looked" once per
                 // Update, which is once per step, so a frame that runs two catch-up steps would
