+++ src/CutTheRopeDX.Desktop.Tests/CutTheRopeDX.Desktop.Tests.csproj
@@ -6,26 +6,15 @@
     <OutputType>Exe</OutputType>
   </PropertyGroup>
 
-  <!--
-    Mirrors the constant the Desktop project defines, so a test that depends on which MonoGame is
-    behind it can say so at compile time. Deriving that from the framework at run time instead
-    would mean a test detecting the absence of the very thing it exists to guard, and passing
-    because of it.
-  -->
-  <PropertyGroup Condition="'$(GraphicsBackend)' == 'GL'">
-    <DefineConstants>$(DefineConstants);MONOGAME_DESKTOPGL</DefineConstants>
-  </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="xunit.v3" />
   </ItemGroup>
 
   <!--
-    Desktop-facing tests: these exercise genuinely Desktop/MonoGame-bound implementation details
-    (quad batching/blend state tied to XNA Rectangle/Matrix/vertex types, the MonoGame Song
-    completion workaround, and FontStashFont's composite-target rotation) that CutTheRopeDX.Tests
-    can no longer compile since it references Core only. Core comes in transitively through
-    Desktop's own ProjectReference.
+    Desktop-facing tests: these exercise the host's own composition - video output, the FFmpeg
+    library resolver, and the update checker - which CutTheRopeDX.Tests cannot reach because it
+    references Core only. Core comes in transitively through Desktop's own ProjectReference.
   -->
   <ItemGroup>
     <ProjectReference Include="..\CutTheRopeDX.Desktop\CutTheRopeDX.Desktop.csproj" />
