+++ src/CutTheRopeDX.Desktop/CutTheRopeDX.Desktop.csproj
@@ -68,29 +68,8 @@
     </BundleResource>
   </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="FontStashSharp.MonoGame" />
-  </ItemGroup>
 
-  <!--
-    Vulkan through MonoGame's native runtime. The framework assembly is backend-agnostic; the
-    MonoGame.Runtime.* packages each contribute the mgruntime native library for one platform.
-  -->
-  <ItemGroup Condition="'$(GraphicsBackend)' == 'VK'">
-    <PackageReference Include="MonoGame.Framework.Native" />
-    <PackageReference Include="MonoGame.Runtime.Windows.Vulkan" />
-    <PackageReference Include="MonoGame.Runtime.Mac.Vulkan" />
-    <PackageReference Include="MonoGame.Runtime.Linux.Vulkan" />
-  </ItemGroup>
 
-  <!--
-    OpenGL, for machines with no usable Vulkan driver. Covers GPUs far older than Vulkan reaches
-    (Vulkan on Intel starts at Skylake; DesktopGL runs on HD 4000 and older) and keeps the Vorbis audio
-    the Vulkan build already uses, so the built content is shared rather than duplicated.
-  -->
-  <ItemGroup Condition="'$(GraphicsBackend)' == 'GL'">
-    <PackageReference Include="MonoGame.Framework.DesktopGL" />
-  </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="../CutTheRopeDX.Core/CutTheRopeDX.Core.csproj" />
@@ -105,16 +84,6 @@
   <ItemGroup Condition="'$(TargetFramework)' == 'net10.0'">
     <PackageReference Include="FFmpeg.AutoGen" />
   </ItemGroup>
-
-  <PropertyGroup Condition="'$(GraphicsBackend)' == 'VK'">
-    <MonoGamePlatform>DesktopVK</MonoGamePlatform>
-    <DefineConstants>$(DefineConstants);MONOGAME_DESKTOPVK</DefineConstants>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(GraphicsBackend)' == 'GL'">
-    <MonoGamePlatform>DesktopGL</MonoGamePlatform>
-    <DefineConstants>$(DefineConstants);MONOGAME_DESKTOPGL</DefineConstants>
-  </PropertyGroup>
 
   <!-- AVFoundation for macOS Tahoe (26) -->
   <PropertyGroup Condition="'$(TargetFramework)' == 'net10.0-macos'">
@@ -146,18 +115,7 @@
     without it. Every other TFM keeps the ordinary CopyToOutputDirectory/CopyToPublishDirectory
     behavior, which already places it next to the executable.
   -->
-  <ItemGroup Condition="'$(TargetFramework)' != 'net10.0-macos'">
-    <None Include="../../alsoft.ini">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </None>
-  </ItemGroup>
 
-  <ItemGroup Condition="'$(TargetFramework)' == 'net10.0-macos'">
-    <BundleResource Include="../../alsoft.ini">
-      <Link>alsoft.ini</Link>
-    </BundleResource>
-  </ItemGroup>
 
   <!--
     macOS: codesign dylibs and the app bundle after build.
