+++ src/CutTheRopeDX.Tests/CutTheRopeDX.Tests.csproj
@@ -6,14 +6,6 @@
     <OutputType>Exe</OutputType>
   </PropertyGroup>
 
-  <!--
-    Mirrors the constant the game project defines, so a test that depends on which MonoGame is behind it
-    can say so at compile time. Deriving that from the framework at run time instead would mean a test
-    detecting the absence of the very thing it exists to guard, and passing because of it.
-  -->
-  <PropertyGroup Condition="'$(GraphicsBackend)' == 'GL'">
-    <DefineConstants>$(DefineConstants);MONOGAME_DESKTOPGL</DefineConstants>
-  </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="xunit.v3" />
